backend/requirements.txt
absl-py==2.1.0
aiohttp==3.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
boto3==1.36.19
botocore==1.36.19
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
darkdetect==0.7.1
decorator==5.1.1
distlib==0.3.9
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.116.0
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
flatbuffers==25.1.24
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
github==1.2.7
GitPython==3.1.44
grpcio==1.70.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.24.7
idna==3.7
ipython==8.32.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
playsound==1.2.2
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
pybboxes==0.1.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.9.3.3
pyqtdarktheme==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.54
s3transfer==0.11.2
sahi==0.11.20
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.29.2
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.18.3
yolov5==7.0.11
